2단계 - 로또(자동) https://edu.nextstep.camp/s/MNii0Puk/ls/uYOLrcvU - (report) 통계 도메인 설계 (통 구매금액, 총 수익, 수익률, 등수별 당첨수)

+++ src/main/java/lotto/domain/Lotto.java
@@ -8,7 +8,7 @@ import java.util.Map;
 public class Lotto {
     private Integer price;
     private Integer[] winningNumber;
-    private final Map<Rank, Integer> prizeMap = new HashMap<>();
+    private final Map<Rank, Long> prizeMap = new HashMap<>();
     private LottoConstraint constraint;
 
     public void setPrice(Integer price) {
@@ -27,11 +27,11 @@ public class Lotto {
         return winningNumber;
     }
 
-    public Map<Rank, Integer> getPrizeMap() {
+    public Map<Rank, Long> getPrizeMap() {
         return prizeMap;
     }
 
-    public void addToPrizeMap(Rank rank, Integer prize) {
+    public void addToPrizeMap(Rank rank, Long prize) {
         prizeMap.put(rank, prize);
     }
